Change email

Change email

Usage

frbs_change_email(id_token, email)

Arguments

id_token

String. A Firebase Auth ID token for the user.

email

String. User's new email.

Value

A named list with the following items:

localId: The uid of the current user.

email: User's email address.

passwordHash: Hash version of password.

providerUserInfo: A named list of of all linked provider objects which contain "providerId" and "federatedId".

idToken: New Firebase Auth ID token for user.

refreshToken: A Firebase Auth refresh token.

expiresIn: string The number of seconds in which the ID token expires.

error:

NULL if no error code in response

A list of 2 if response was an error:

code: Error code

message: Error message

Details

DISCLAIMER: Changing a users's email requires that you disable email enumeration protection for your firebase project. This is NOT recommended.

Learn about email enumeration protection

Visit Firebase Auth REST API docs for more details

Examples

if (FALSE) { # \dontrun{
# first sign in user and get the 'id_token':
user <- frbs_sign_in(email = "user@gmail.com", password = "password")
id_token <- user$idToken

# change email:
response <- frbs_change_email(
  id_token = id_token,
  email = "new.email@mail.com"
)
response
} # }