Exchange custom token for an ID and refresh token

Exchanges a custom Auth token for an ID and refresh token

Usage

frbs_exchange_custom_token(token)

Arguments

token

String. A Firebase Auth custom token from which to create an ID and refresh token pair

Value

A named list with the following items:

idToken: A Firebase Auth ID token generated from the provided custom token.

refreshToken: A Firebase Auth refresh token generated from the provided custom token.

expiresIn: The number of seconds in which the ID token expires.

error:

NULL if no error code in response

A list of 2 if response was an error:

code: Error code

message: Error message

Details

Visit Firebase Auth REST API docs for more details