Update user profile

Update a user's profile (display name / photo URL).

Usage

frbs_update_profile(
  id_token,
  display_name = NULL,
  photo_url = NULL,
  delete_attribute = NULL
)

Arguments

id_token

String. A Firebase Auth ID token for the user.

display_name

String. User's new display name. Defaults to NULL.

photo_url

String. User's new photo url. Defaults to NULL.

delete_attribute

Character vector of attributes to delete. Either "DISPLAY_NAME" or "PHOTO_URL". This will nullify these values. Defaults to NULL.

Value

A named list with the following items:

localId: The uid of the current user.

email: User's email address.

displayName: User's new display name.

photoUrl: User's new photo url.

passwordHash: Hash version of password.

providerUserInfo: A named list of of all linked provider objects which contain "providerId" and "federatedId".

idToken: New Firebase Auth ID token for user.

refreshToken: A Firebase Auth refresh token.

expiresIn: string The number of seconds in which the ID token expires.

error:

NULL if no error code in response

A list of 2 if response was an error:

code: Error code

message: Error message

Details

Visit Firebase Auth REST API docs for more details

Examples

if (FALSE) {
frbs_update_profile(
  id_token = "id-token-goes-here",
  display_name = "new-user-display-name",
  photo_url = "url-to-user-photo"
)

# to delete the display name attribute:
frbs_update_profile(
  delete_attribute = "DISPLAY_NAME"
)
}